Fix starting after stop has been called.
Previously, when the server was stopped, a null value was assigned to the
variable holding the socket. However, a new socket wasn't created during the
start call (that was done only when creating a server).

After this commit, a new socket is created during starting the server so that
a server can be stopped and started several times without re-creating it.

Added also an optional callback argument to start and stop that can be used
if an application needs to track more exactly when the underlaying async
operations are done.